+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Menu Component
 - Cards Component
 - Labels Component
+- Breadcrumb Component
 
 ### Changed