+++ CHANGELOG.md
@@ -66,3 +66,9 @@ Initial release of the Design System Package
 ### Added
 
 - Card/Caption Typography
+
+## [v1.18.0] - 2022-01-20
+
+### Added
+
+- Updated tailwind to version 3